Revert changes from refactor commit (#154)
