+++ src/Services/dataDownload/dataExportService.js
@@ -1,11 +1,7 @@
 import DownloadService from './downloadService';
 import apiService from '../../api/api';
 import datasetMetadataToDownloadFormat from './datasetMetadataToDownloadFormat';
-import InputProcessorFactory from './processors/InputProcessorFactory';
-import ExportFormatDecider from './ExportFormatDecider';
-import OutputGeneratorFactory from './generators/OutputGeneratorFactory';
-import ExportErrorHandler from './ExportErrorHandler';
-import ArrayBufferProcessor from './processors/ArrayBufferProcessor';
+import Papa from 'papaparse';
 /**
  * Unified data export service for handling all export operations
  */
@@ -28,62 +24,159 @@ class DataExportService {
     variableName = null,
     forceZip = false,
   }) {
-    try {
-      // Process input data using appropriate processor
-      const processedData = InputProcessorFactory.process(data);
+    // Handle ArrayBuffer input (new capability for large datasets)
+    if (data instanceof ArrayBuffer) {
+      const bufferSize = data.byteLength;
 
-      // Handle large buffers that require special processing
-      if (
-        processedData.requiresSpecialHandling &&
-        processedData.type === 'large_buffer'
-      ) {
-        return ArrayBufferProcessor.handleLargeBuffer(
-          processedData.data,
+      if (!DataExportService.canSafelyDecodeBuffer(bufferSize)) {
+        return DataExportService.downloadLargeBufferAsCSV(
+          data,
           datasetName,
           metadata,
         );
       }
 
-      // Determine export format based on data characteristics
-      const format = ExportFormatDecider.determineFormat(processedData, {
-        forceZip,
-      });
+      // Convert to string for normal processing
+      try {
+        data = new TextDecoder().decode(data);
 
-      // Generate filename
-      const baseFilename = variableName
-        ? `${datasetName}_${variableName}_${DownloadService.formatDateForFilename()}`
-        : `${datasetName}_${DownloadService.formatDateForFilename()}`;
+        // Validate conversion was successful
+        if (!data || data.length === 0) {
+          throw new Error('TextDecoder returned empty string');
+        }
+      } catch (decodingError) {
+        console.error(
+          'TextDecoder failed, falling back to direct CSV download',
+          {
+            error: decodingError.message,
+            bufferSize: bufferSize,
+          },
+        );
 
-      // Generate and download using appropriate generator
-      await OutputGeneratorFactory.generate(
-        format,
-        processedData,
-        metadata,
-        datasetName,
-        baseFilename,
-      );
+        return DataExportService.downloadLargeBufferAsCSV(
+          data,
+          datasetName,
+          metadata,
+        );
+      }
+    }
 
-      ExportErrorHandler.logOperation('export', {
-        datasetName,
-        variableName,
-        format,
-        dataType: processedData.type,
-      });
-    } catch (error) {
-      // Use standardized error handling with fallback strategies
-      const processedData = InputProcessorFactory.process(data);
-      const baseFilename = variableName
-        ? `${datasetName}_${variableName}_${DownloadService.formatDateForFilename()}`
-        : `${datasetName}_${DownloadService.formatDateForFilename()}`;
+    // Normalize data to consistent format
+    let normalizedData;
+    let csvData;
 
-      await ExportErrorHandler.handleExportError(
-        error,
-        processedData,
-        metadata,
-        datasetName,
-        baseFilename,
+    if (typeof data === 'string') {
+      // Data is CSV string
+      csvData = data;
+      normalizedData = DataExportService.parseCSVToJSON(data);
+    } else if (Array.isArray(data)) {
+      // Data is JSON array
+      normalizedData = DataExportService.normalizeVisualizationData(data);
+      csvData = DataExportService.convertVisualizationDataToCSV(data);
+    } else {
+      throw new Error(
+        'Data must be either a CSV string, JSON array, or ArrayBuffer',
       );
     }
+
+    // Determine if data is too large for Excel
+    const shouldUseZip =
+      forceZip || DataExportService.shouldUseZipFormat(normalizedData, csvData);
+
+    // Generate filename components
+    const baseFilename = variableName
+      ? `${datasetName}_${variableName}_${DownloadService.formatDateForFilename()}`
+      : `${datasetName}_${DownloadService.formatDateForFilename()}`;
+
+    if (shouldUseZip) {
+      // Export as ZIP with CSV data and Excel metadata
+      const metadataWorkbook = DownloadService.createExcelWorkbook(
+        DownloadService.createMetadataSheets(metadata),
+      );
+      const metadataBuffer = DownloadService.workbookToBuffer(metadataWorkbook);
+
+      const files = [
+        {
+          filename: `${datasetName}_data.csv`,
+          content: csvData,
+        },
+        {
+          filename: `${datasetName}_metadata.xlsx`,
+          content: metadataBuffer,
+        },
+      ];
+
+      await DownloadService.downloadZip(files, baseFilename);
+    } else {
+      // Export as single Excel file with data and metadata sheets
+      const sheets = [
+        {
+          name: 'Data',
+          data: normalizedData,
+        },
+        ...DownloadService.createMetadataSheets(metadata),
+      ];
+
+      try {
+        DownloadService.downloadExcel(sheets, baseFilename);
+      } catch (error) {
+        console.warn(
+          'Excel download failed, falling back to ZIP format:',
+          error.message,
+        );
+
+        // Fallback to ZIP format
+        const metadataWorkbook = DownloadService.createExcelWorkbook(
+          DownloadService.createMetadataSheets(metadata),
+        );
+        const metadataBuffer =
+          DownloadService.workbookToBuffer(metadataWorkbook);
+
+        const files = [
+          {
+            filename: `${datasetName}_data.csv`,
+            content: csvData,
+          },
+          {
+            filename: `${datasetName}_metadata.xlsx`,
+            content: metadataBuffer,
+          },
+        ];
+
+        await DownloadService.downloadZip(files, baseFilename);
+      }
+    }
+  }
+
+  /**
+   * Determines if ZIP format should be used based on data size
+   * @param {Array} normalizedData - Data as JSON array
+   * @param {string} csvData - Data as CSV string
+   * @returns {boolean} True if ZIP format should be used
+   */
+  static shouldUseZipFormat(normalizedData, csvData) {
+    // Excel practical limits:
+    // - ~1 million rows (1,048,576 actual limit)
+    // - Memory considerations for large datasets
+    // - File size considerations
+
+    const MAX_EXCEL_ROWS = 500000; // Conservative limit for performance
+    const MAX_CSV_SIZE_MB = 25; // 50MB CSV size limit for Excel export
+
+    // Check row count
+    if (normalizedData && normalizedData.length > MAX_EXCEL_ROWS) {
+      return true;
+    }
+
+    // Check approximate file size (CSV string length as rough estimate)
+    if (csvData) {
+      const csvSizeMB = new Blob([csvData]).size / (1024 * 1024);
+      if (csvSizeMB > MAX_CSV_SIZE_MB) {
+        return true;
+      }
+    }
+
+    return false;
   }
 
   /**
@@ -119,6 +212,117 @@ class DataExportService {
       console.error('Error fetching metadata:', error);
       throw error;
     }
+  }
+
+  /**
+   * Convert visualization data to CSV format
+   * @param {Array} data - Visualization data array
+   * @returns {string} CSV formatted string
+   */
+  static convertVisualizationDataToCSV(data) {
+    if (!data || data.length === 0) {
+      return '';
+    }
+
+    // Normalize data first to ensure consistent structure
+    const normalizedData = DataExportService.normalizeVisualizationData(data);
+
+    // Determine column order based on data type
+    const columns =
+      DataExportService.determineVisualizationColumns(normalizedData);
+
+    return DownloadService.jsonToCSV(normalizedData, columns);
+  }
+
+  /**
+   * Normalize visualization data to consistent format
+   * @param {Array} data - Raw visualization data
+   * @returns {Array} Normalized data array
+   */
+  static normalizeVisualizationData(data) {
+    // Handle different visualization data formats
+    return data.map((row) => {
+      const normalized = { ...row };
+
+      // Ensure consistent date formatting
+      if (row.time) {
+        normalized.time = DataExportService.formatDate(row.time);
+      }
+
+      // Ensure numeric values are properly formatted
+      Object.keys(normalized).forEach((key) => {
+        if (
+          typeof normalized[key] === 'number' &&
+          !Number.isInteger(normalized[key])
+        ) {
+          normalized[key] = DataExportService.formatFloat(normalized[key]);
+        }
+      });
+
+      return normalized;
+    });
+  }
+
+  /**
+   * Determine column order for visualization data
+   * @param {Array} data - Normalized data array
+   * @returns {Array<string>} Ordered column names
+   */
+  static determineVisualizationColumns(data) {
+    if (!data || data.length === 0) {
+      return [];
+    }
+
+    const firstRow = data[0];
+    const columns = Object.keys(firstRow);
+
+    // Define preferred column order
+    const orderedColumns = [];
+    const preferredOrder = ['time', 'lat', 'lon', 'depth'];
+
+    // Add preferred columns in order if they exist
+    preferredOrder.forEach((col) => {
+      if (columns.includes(col)) {
+        orderedColumns.push(col);
+      }
+    });
+
+    // Add remaining columns
+    columns.forEach((col) => {
+      if (!orderedColumns.includes(col)) {
+        orderedColumns.push(col);
+      }
+    });
+
+    return orderedColumns;
+  }
+
+  /**
+   * Format date for export
+   * @param {string|Date} date - Date to format
+   * @returns {string} Formatted date string
+   */
+  static formatDate(date) {
+    if (!date) {
+      return '';
+    }
+
+    const dateObj = date instanceof Date ? date : new Date(date);
+    if (isNaN(dateObj.getTime())) {
+      return String(date);
+    }
+
+    return dateObj.toISOString();
+  }
+
+  /**
+   * Format float values with appropriate precision
+   * @param {number} value - Float value to format
+   * @param {number} precision - Number of decimal places (default: 6)
+   * @returns {number} Formatted float
+   */
+  static formatFloat(value, precision = 6) {
+    return parseFloat(value.toFixed(precision));
   }
 
   /**
@@ -158,6 +362,114 @@ class DataExportService {
     };
 
     return filteredMetadata;
+  }
+
+  /**
+   * Check if buffer can be safely decoded by TextDecoder
+   * @param {number} bufferSize - Size in bytes
+   * @returns {boolean} True if safe to decode
+   */
+  static canSafelyDecodeBuffer(bufferSize) {
+    const MAX_TEXTDECODER_BUFFER = 100 * 1024 * 1024; // 100MB safe limit for TextDecoder
+    return bufferSize <= MAX_TEXTDECODER_BUFFER;
+  }
+
+  /**
+   * Download large buffer as ZIP with CSV data and Excel metadata
+   * @param {ArrayBuffer} buffer - Raw data buffer from API
+   * @param {string} datasetName - Name of the dataset
+   * @param {Object} metadata - Dataset metadata for Excel generation
+   * @returns {Promise<void>}
+   */
+  static async downloadLargeBufferAsCSV(buffer, datasetName, metadata) {
+    try {
+      const baseFilename = `${datasetName}_${DownloadService.formatDateForFilename()}`;
+
+      // Create Excel metadata file
+      const metadataWorkbook = DownloadService.createExcelWorkbook(
+        DownloadService.createMetadataSheets(metadata),
+      );
+      const metadataBuffer = DownloadService.workbookToBuffer(metadataWorkbook);
+
+      // Create ZIP with CSV data and Excel metadata
+      const files = [
+        {
+          filename: `${datasetName}_data.csv`,
+          content: buffer, // ArrayBuffer content
+        },
+        {
+          filename: `${datasetName}_metadata.xlsx`,
+          content: metadataBuffer,
+        },
+      ];
+
+      await DownloadService.downloadZip(files, baseFilename);
+
+      console.info('Successfully downloaded large dataset as ZIP', {
+        datasetName,
+        sizeMB: (buffer.byteLength / (1024 * 1024)).toFixed(1),
+      });
+    } catch (error) {
+      console.error('Failed to download large buffer as ZIP', {
+        error: error.message,
+        datasetName,
+      });
+      throw new Error(`Failed to download dataset: ${error.message}`);
+    }
+  }
+
+  /**
+   * Parse CSV string data to JSON array
+   * @param {string} csvData - CSV formatted string data
+   * @returns {Array<Object>} Array of objects with headers as keys
+   * @throws {Error} If CSV data is invalid or empty
+   */
+  static parseCSVToJSON(csvData) {
+    if (!csvData || typeof csvData !== 'string') {
+      throw new Error('Invalid CSV data: must be a non-empty string');
+    }
+
+    const trimmedData = csvData.trim();
+    if (!trimmedData) {
+      throw new Error('CSV data is empty');
+    }
+
+    try {
+      // Use Papa Parse for robust CSV parsing that handles escaping properly
+      const parseResult = Papa.parse(trimmedData, {
+        header: true,
+        skipEmptyLines: true,
+        transformHeader: (header) => header.trim(),
+        transform: (value) => value.trim(),
+      });
+
+      if (parseResult.errors && parseResult.errors.length > 0) {
+        // Log the first parsing error for debugging
+        const firstError = parseResult.errors[0];
+        throw new Error(
+          `CSV parsing error at row ${firstError.row}: ${firstError.message}`,
+        );
+      }
+
+      if (!parseResult.data || parseResult.data.length === 0) {
+        throw new Error(
+          'CSV data must have at least a header row and one data row',
+        );
+      }
+
+      return parseResult.data;
+    } catch (error) {
+      // If it's already our custom error, re-throw it
+      if (
+        error.message.includes('CSV parsing error') ||
+        error.message.includes('CSV data must have')
+      ) {
+        throw error;
+      }
+
+      // Otherwise, wrap the Papa Parse error with more context
+      throw new Error(`Failed to parse CSV data: ${error.message}`);
+    }
   }
 }
